Fixed issues with CAPTCHA solving!

+++ app.js
@@ -2,10 +2,13 @@ import fetch from "node-fetch";
 import boidInfo from "./data/data.js";
 import colors from "colors";
 import promptSync from "prompt-sync";
+import { createSpinner } from "nanospinner";
 import { createWorker, OEM } from "tesseract.js";
 
 const prompt = promptSync();
 const url = "https://iporesult.cdsc.com.np/";
+
+const sleep = (ms = 2000) => new Promise((r) => setTimeout(r, ms));
 
 const getData = () => {
   return fetch(url + "result/companyShares/fileUploaded")
@@ -14,6 +17,26 @@ const getData = () => {
       console.error("Error:", err);
       process.exit(1);
     });
+};
+
+const postData = async (userID, v, userCaptcha, captchaIdentifier) => {
+  try {
+    const res = await fetch(url + "result/result/check", {
+      headers: {
+        "Content-Type": "application/json",
+      },
+      body: JSON.stringify({
+        companyShareId: userID,
+        boid: v,
+        userCaptcha: userCaptcha,
+        captchaIdentifier: captchaIdentifier,
+      }),
+      method: "POST",
+    });
+    return await res.json();
+  } catch (error) {
+    console.log("Something went wrong!".red.bold.underline);
+  }
 };
 
 const workers = [];
@@ -42,7 +65,7 @@ getData()
     const checking = details.filter((a) => a.id == userID)[0];
     console.log("Checking Result of", `${checking.name}`.cyan.underline, "\n");
 
-    const result = async (worker, v) => {
+    const result = async (worker, v, captchaSpinner) => {
       let userCaptcha,
         captcha,
         captchaIdentifier = null;
@@ -55,7 +78,7 @@ getData()
         tessedit_char_whitelist: "0123456789",
       });
 
-      while (!/^(\d{5})$/.test(userCaptcha)) {
+      while (true) {
         let data = await getData();
         ({ captcha, captchaIdentifier } = data.body.captchaData);
         let captchaImg = Buffer.from(captcha, "base64");
@@ -63,31 +86,25 @@ getData()
           data: { text },
         } = await worker.recognize(captchaImg);
         userCaptcha = text.replace(/[^0-9]/g, "");
-      }
-      await worker.terminate();
 
-      return fetch(url + "result/result/check", {
-        headers: {
-          "Content-Type": "application/json",
-        },
-        body: JSON.stringify({
-          companyShareId: userID,
-          boid: v,
-          userCaptcha: userCaptcha,
-          captchaIdentifier: captchaIdentifier,
-        }),
-        method: "POST",
-      })
-        .then((res) => res.json())
-        .catch((error) => {
-          console.log("Something went wrong!".red.bold.underline);
-        });
+        let results = await postData(userID, v, userCaptcha, captchaIdentifier);
+        if (
+          results.message &&
+          results.message !== "Invalid Captcha Provided. Please try again "
+        ) {
+          await worker.terminate();
+          captchaSpinner.success({ text: "Completed!" });
+          return results;
+        }
+      }
     };
 
     let workerIndex = 0;
-    boidInfo.map((user) => {
+    boidInfo.map(async (user) => {
       if (user.boid) {
-        result(workers[workerIndex++], user.boid).then((r) =>
+        const captchaSpinner = createSpinner("Solving CAPTCHA...").start();
+        await sleep();
+        result(workers[workerIndex++], user.boid, captchaSpinner).then((r) =>
           typeof r === "undefined"
             ? console.log(
                 `${user.name} => Possible Error: Incorrect BOID`.yellow + "\n"
